[chore] 복약봉투 인식 수정
